add option to explicitly override file when creating writer

+++ rs-utils/src/fs.rs
@@ -415,7 +415,11 @@ pub fn create_file_reader(file_path: &str) -> Result<BufReader<File>> {
     Ok(data_reader)
 }
 
-pub fn create_file_writer(file_path: &str) -> Result<impl Write> {
+pub fn create_file_writer(file_path: &str, overwrite: bool) -> Result<impl Write> {
+    if !overwrite && file_exists(file_path)? {
+        bail!("File {file_path} already exists");
+    }
+
     let new_file = File::create(file_path)?;
     let data_writer = BufWriter::new(new_file);
